% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{american_optprice}
\alias{american_optprice}
\title{American Call Option Pricing}
\usage{
american_optprice(S_0, S_t, tm, vol, dy = 0, r = 0.0409)
}
\arguments{
\item{S_0}{Underlying asset price}

\item{S_t}{Strike price}

\item{vol}{Implied volatility}

\item{dy}{Dividend yield}

\item{r}{Annual risk-free interest rate}

\item{t}{Time to maturity (days)}
}
\value{
Vector of option values
}
\description{
Calculates American call option values using RQuantLib
}
