% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{american_optprice_returns}
\alias{american_optprice_returns}
\title{American Option Returns Calculator}
\usage{
american_optprice_returns(data, K, tm, vol_0 = NA, vol_t = NA)
}
\arguments{
\item{data}{Data frame with time (t) and returns (R) columns}

\item{K}{Strike price}

\item{tm}{Time to maturity (days)}
}
\value{
Data frame with added option columns (opt_entry, opt_exit, opt_R)
}
\description{
Calculates option entry, exit prices and returns for given data
}
