% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserved_words.R
\docType{data}
\name{constants}
\alias{constants}
\title{Constants}
\format{
An object of class \code{list} of length 17.
}
\usage{
constants
}
\description{
A list of constants used throughout the framework to control
the simulation. These are considered global constants.
}
\section{Log Levels}{

\itemize{
\item \code{ERROR}: Error log level
\item \code{WARNING}: Warning log level
\item \code{INFO}: Info log level
\item \code{DEBUG}: Debug log level
\item \code{LOG_LEVELS}: Available log levels
}
}

\section{Time Frames}{

\itemize{
\item \code{M1}: Time frame: M1
\item \code{M5}: Time frame: M5
\item \code{M15}: Time frame: M15
\item \code{M30}: Time frame: M30
\item \code{H1}: Time frame: H1
\item \code{H4}: Time frame: H4
\item \code{D1}: Time frame: D1
\item \code{W1}: Time frame: W1
\item \code{MN1}: Time frame: MN1
\item \code{Q}: Time frame: Q
\item \code{Y}: Time frame: Y
\item \code{TIMEFRAMES}: Available time frames
}
}

\keyword{datasets}
