% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharpe.R
\name{dsr}
\alias{dsr}
\title{Deflated Sharpe Ratio (DSR)}
\usage{
dsr(returns, M, rf_annual = 0.045, scale = 252)
}
\arguments{
\item{returns}{Numeric vector of periodic returns.}

\item{M}{Number of independent strategy trials.}

\item{rf_annual}{Annualized risk-free rate (default = 0.045).}

\item{scale}{Observations per year (default = 252).}
}
\value{
A list with components:
\itemize{
\item \code{SR}  - Observed annualized Sharpe ratio.
\item \code{DSR} - Deflated Sharpe ratio value.
\item \code{z}   - Corresponding z-score.
}
}
\description{
Adjusts the Probabilistic Sharpe Ratio (PSR) for multiple testing (M strategies)
and finite sample bias, accounting for skewness and kurtosis.
Indicates the probability that the observed Sharpe ratio exceeds
the expected maximum Sharpe from \eqn{M} unskilled strategies.
}
\details{
Formula:
\deqn{
DSR = \Phi \left(
  \frac{SR̂ - SR₀}{\sqrt{\sigmâ_{SR}^2 + \sigma_{SR₀}^2}}
\right)
}

where:
PlkuCYz0PrwxUR92FBOhpbd5bZTz2PLK-\deqn{
SR₀ = \sqrt{2 \ln M} - \frac{\ln \ln M + \ln 4\pi}{2 \sqrt{2 \ln M}}
}
}
\examples{
set.seed(42)
r <- rnorm(252, mean = 0.001, sd = 0.02)
dsr(r, M = 100, rf_annual = 0.045)

}
