% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetl.R
\name{Fetl}
\alias{Fetl}
\title{Datasource Postgres}
\description{
A singleton, half-duplex/simplex Postgres connector to fetch
data quotes from a Postgres database.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Fetl-new}{\code{Fetl$new()}}
\item \href{#method-Fetl-connect}{\code{Fetl$connect()}}
\item \href{#method-Fetl-disconnect}{\code{Fetl$disconnect()}}
\item \href{#method-Fetl-send_query}{\code{Fetl$send_query()}}
\item \href{#method-Fetl-execute_query}{\code{Fetl$execute_query()}}
\item \href{#method-Fetl-fsg}{\code{Fetl$fsg()}}
\item \href{#method-Fetl-build_fsg}{\code{Fetl$build_fsg()}}
\item \href{#method-Fetl-quotes}{\code{Fetl$quotes()}}
\item \href{#method-Fetl-build_quotes_query}{\code{Fetl$build_quotes_query()}}
\item \href{#method-Fetl-stock_forward_mass}{\code{Fetl$stock_forward_mass()}}
\item \href{#method-Fetl-clone}{\code{Fetl$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fetl-new"></a>}}
\if{latex}{\out{\hypertarget{method-Fetl-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fetl$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fetl-connect"></a>}}
\if{latex}{\out{\hypertarget{method-Fetl-connect}{}}}
\subsection{Method \code{connect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fetl$connect(force_reconnect = FALSE, verbose = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fetl-disconnect"></a>}}
\if{latex}{\out{\hypertarget{method-Fetl-disconnect}{}}}
\subsection{Method \code{disconnect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fetl$disconnect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fetl-send_query"></a>}}
\if{latex}{\out{\hypertarget{method-Fetl-send_query}{}}}
\subsection{Method \code{send_query()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fetl$send_query(query, timeseries = FALSE, xts = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fetl-execute_query"></a>}}
\if{latex}{\out{\hypertarget{method-Fetl-execute_query}{}}}
\subsection{Method \code{execute_query()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fetl$execute_query(query)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fetl-fsg"></a>}}
\if{latex}{\out{\hypertarget{method-Fetl-fsg}{}}}
\subsection{Method \code{fsg()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fetl$fsg(
  n = 5,
  allow_partial_results = FALSE,
  start_date = NULL,
  end_date = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fetl-build_fsg"></a>}}
\if{latex}{\out{\hypertarget{method-Fetl-build_fsg}{}}}
\subsection{Method \code{build_fsg()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fetl$build_fsg(n, allow_partial_results, start_date, end_date)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fetl-quotes"></a>}}
\if{latex}{\out{\hypertarget{method-Fetl-quotes}{}}}
\subsection{Method \code{quotes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fetl$quotes(
  symbols,
  from = NULL,
  to = NULL,
  exchange = NULL,
  sector = NULL,
  industry = NULL,
  country = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fetl-build_quotes_query"></a>}}
\if{latex}{\out{\hypertarget{method-Fetl-build_quotes_query}{}}}
\subsection{Method \code{build_quotes_query()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fetl$build_quotes_query(symbols, from, to, exchange, sector, industry, country)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fetl-stock_forward_mass"></a>}}
\if{latex}{\out{\hypertarget{method-Fetl-stock_forward_mass}{}}}
\subsection{Method \code{stock_forward_mass()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fetl$stock_forward_mass(from_date, to_date = today())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fetl-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Fetl-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fetl$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
