% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lock.R
\name{lock_all}
\alias{lock_all}
\title{Lock variable bindings}
\usage{
lock_all(..., env = .GlobalEnv)
}
\arguments{
\item{...}{Variable names to lock (unquoted)}

\item{env}{Environment containing the variables (default: .GlobalEnv)}
}
\value{
NULL (invisibly)
}
\description{
Lock variable bindings
}
