% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserved_words.R
\docType{data}
\name{M15}
\alias{M15}
\title{Timeframes: every quarter hour}
\format{
An object of class \code{character} of length 1.
}
\usage{
M15
}
\description{
Timeframes: every quarter hour
}
\keyword{datasets}
