% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserved_words.R
\docType{data}
\name{M30}
\alias{M30}
\title{Timeframes: every half hour}
\format{
An object of class \code{character} of length 1.
}
\usage{
M30
}
\description{
Timeframes: every half hour
}
\keyword{datasets}
