% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_optim.R
\name{options_optim_maximal}
\alias{options_optim_maximal}
\title{Find Optimal Parameters}
\usage{
options_optim_maximal(surface_grid)
}
\arguments{
\item{surface_grid}{Data frame from options_optim_surface_grid}
}
\value{
Single-row data frame with optimal K, tm, and result values
}
\description{
Extracts parameters corresponding to maximum optimization result
}
