% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_optim.R
\name{options_optim_surface_matrix}
\alias{options_optim_surface_matrix}
\title{Options Optimization Surface Matrix}
\usage{
options_optim_surface_matrix(surface_grid, K_length, tm_values)
}
\arguments{
\item{surface_grid}{Data frame from options_optim_surface_grid}

\item{K_length}{Number of strike price values}

\item{tm_values}{Number of maturity values}
}
\value{
Matrix of optimization results with tm rows and K columns
}
\description{
Converts optimization surface grid to matrix format for visualization
}
