% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concurrency_plots.R
\name{plot_concurrency_distribution}
\alias{plot_concurrency_distribution}
\title{Plot concurrency distribution}
\usage{
plot_concurrency_distribution(df_dates, plot = TRUE)
}
\arguments{
\item{df_dates}{Input data frame.}

\item{plot}{Logical; if TRUE prints plot.}
}
\value{
List with data and ggplot.
}
\description{
Histogram of number of concurrent trades per position.
}
