% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concurrency_plots.R
\name{plot_concurrency_over_time}
\alias{plot_concurrency_over_time}
\title{Plot trade concurrency over time}
\usage{
plot_concurrency_over_time(df_dates, plot = TRUE)
}
\arguments{
\item{df_dates}{Input data frame.}

\item{plot}{Logical; if TRUE prints plot.}
}
\value{
List with data and ggplot.
}
\description{
Bar plot of number of active trades per day.
}
