% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concurrency_plots.R
\name{plot_concurrency_punchcard}
\alias{plot_concurrency_punchcard}
\title{Plot weekly punchcard}
\usage{
plot_concurrency_punchcard(df_dates, plot = TRUE)
}
\arguments{
\item{df_dates}{Input data frame.}

\item{plot}{Logical; if TRUE prints plot.}
}
\value{
List with data and ggplot.
}
\description{
GitHub-style weekly concurrency heatmap.
}
