% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharpe.R
\name{psr}
\alias{psr}
\title{Probabilistic Sharpe Ratio (PSR)}
\usage{
psr(returns, sr_benchmark = 0, rf_annual = 0.045, scale = 252)
}
\arguments{
\item{returns}{Numeric vector of periodic returns.}

\item{sr_benchmark}{Sharpe ratio benchmark to test against (default = 0).}

\item{rf_annual}{Annualized risk-free rate (default = 0.045).}

\item{scale}{Observations per year (default = 252).}
}
\value{
A list with components:
\itemize{
\item \code{SR}  - Observed annualized Sharpe ratio.
\item \code{PSR} - Probabilistic Sharpe ratio value.
\item \code{z}   - Corresponding z-score.
}
}
\description{
Estimates the probability that the true Sharpe ratio exceeds
a chosen benchmark \eqn{SR^*}, adjusting for skewness (\eqn{\gamma_3})
and kurtosis (\eqn{\gamma_4}) of returns.
}
\details{
Formula:
\deqn{
PSR(SR^*) = \Phi \left(
  \frac{(SR̂ - SR^*) \sqrt{n - 1}}
       {\sqrt{1 - \gamma_3 SR̂ + ((\gamma_4 - 1)/4) SR̂^2}}
\right)
}
}
\examples{
set.seed(42)
r <- rnorm(252, mean = 0.001, sd = 0.02)
psr(r, sr_benchmark = 0, rf_annual = 0.045)

}
