% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qetl.R
\name{Qetl}
\alias{Qetl}
\title{Datasource Postgres}
\description{
A singleton, half-duplex/simplex Postgres connector to fetch
data quotes from a Postgres database.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Qetl-new}{\code{Qetl$new()}}
\item \href{#method-Qetl-connect}{\code{Qetl$connect()}}
\item \href{#method-Qetl-disconnect}{\code{Qetl$disconnect()}}
\item \href{#method-Qetl-send_query}{\code{Qetl$send_query()}}
\item \href{#method-Qetl-aggregates}{\code{Qetl$aggregates()}}
\item \href{#method-Qetl-query_aggregates}{\code{Qetl$query_aggregates()}}
\item \href{#method-Qetl-univariates}{\code{Qetl$univariates()}}
\item \href{#method-Qetl-query_univariates}{\code{Qetl$query_univariates()}}
\item \href{#method-Qetl-kind}{\code{Qetl$kind()}}
\item \href{#method-Qetl-clone}{\code{Qetl$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Qetl-new"></a>}}
\if{latex}{\out{\hypertarget{method-Qetl-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Qetl$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Qetl-connect"></a>}}
\if{latex}{\out{\hypertarget{method-Qetl-connect}{}}}
\subsection{Method \code{connect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Qetl$connect(force_reconnect = FALSE, verbose = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Qetl-disconnect"></a>}}
\if{latex}{\out{\hypertarget{method-Qetl-disconnect}{}}}
\subsection{Method \code{disconnect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Qetl$disconnect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Qetl-send_query"></a>}}
\if{latex}{\out{\hypertarget{method-Qetl-send_query}{}}}
\subsection{Method \code{send_query()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Qetl$send_query(query, timeseries = FALSE, xts = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Qetl-aggregates"></a>}}
\if{latex}{\out{\hypertarget{method-Qetl-aggregates}{}}}
\subsection{Method \code{aggregates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Qetl$aggregates(ticker, timeframe = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Qetl-query_aggregates"></a>}}
\if{latex}{\out{\hypertarget{method-Qetl-query_aggregates}{}}}
\subsection{Method \code{query_aggregates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Qetl$query_aggregates(ticker, timeframe)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Qetl-univariates"></a>}}
\if{latex}{\out{\hypertarget{method-Qetl-univariates}{}}}
\subsection{Method \code{univariates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Qetl$univariates(ticker, timeframe = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Qetl-query_univariates"></a>}}
\if{latex}{\out{\hypertarget{method-Qetl-query_univariates}{}}}
\subsection{Method \code{query_univariates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Qetl$query_univariates(ticker, timeframe)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Qetl-kind"></a>}}
\if{latex}{\out{\hypertarget{method-Qetl-kind}{}}}
\subsection{Method \code{kind()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Qetl$kind(ticker)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Qetl-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Qetl-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Qetl$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
