% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lock.R
\name{unlock_all}
\alias{unlock_all}
\title{Unlock variable bindings}
\usage{
unlock_all(..., env = .GlobalEnv)
}
\arguments{
\item{...}{Variable names to unlock (unquoted)}

\item{env}{Environment containing the variables (default: .GlobalEnv)}
}
\value{
NULL (invisibly)
}
\description{
Unlock variable bindings
}
