% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserved_words.R
\docType{data}
\name{WARNING}
\alias{WARNING}
\title{Level level: warning}
\format{
An object of class \code{character} of length 1.
}
\usage{
WARNING
}
\description{
Level level: warning
}
\keyword{datasets}
